typing fixed in face of ongoing python 3.6 support
