+++ .pre-commit-config.yaml
@@ -1,3 +1,5 @@
+default_language_version:
+  python: python3.9
 repos:
 -   repo: https://github.com/psf/black
     rev: 22.3.0
