+++ testing/conftest.py
@@ -1,7 +1,9 @@
-import itertools
 import os
 
 import pytest
+
+from .wd_wrapper import WorkDir
+
 
 # 2009-02-13T23:31:30+00:00
 os.environ["SOURCE_DATE_EPOCH"] = "1234567890"
@@ -10,13 +12,15 @@ VERSION_PKGS = ["setuptools", "setuptools_scm"]
 
 
 def pytest_report_header():
-    import pkg_resources
-
+    try:
+        from importlib.metadata import version
+    except ImportError:
+        from importlib_metadata import version
     res = []
     for pkg in VERSION_PKGS:
-        version = pkg_resources.get_distribution(pkg).version
+        pkg_version = version(pkg)
         path = __import__(pkg).__file__
-        res.append(f"{pkg} version {version} from {path!r}")
+        res.append(f"{pkg} version {pkg_version} from {path!r}")
     return res
 
 
@@ -25,72 +29,6 @@ def pytest_addoption(parser):
     group.addoption(
         "--test-legacy", dest="scm_test_virtualenv", default=False, action="store_true"
     )
-
-
-class Wd:
-    commit_command = None
-    signed_commit_command = None
-    add_command = None
-
-    def __repr__(self):
-        return f"<WD {self.cwd}>"
-
-    def __init__(self, cwd):
-        self.cwd = cwd
-        self.__counter = itertools.count()
-
-    def __call__(self, cmd, **kw):
-        if kw:
-            cmd = cmd.format(**kw)
-        from setuptools_scm.utils import do
-
-        return do(cmd, self.cwd)
-
-    def write(self, name, value, **kw):
-        filename = self.cwd / name
-        if kw:
-            value = value.format(**kw)
-        if isinstance(value, bytes):
-            filename.write_bytes(value)
-        else:
-            filename.write_text(value)
-        return filename
-
-    def _reason(self, given_reason):
-        if given_reason is None:
-            return f"number-{next(self.__counter)}"
-        else:
-            return given_reason
-
-    def add_and_commit(self, reason=None, **kwargs):
-        self(self.add_command)
-        self.commit(reason, **kwargs)
-
-    def commit(self, reason=None, signed=False):
-        reason = self._reason(reason)
-        self(
-            self.commit_command if not signed else self.signed_commit_command,
-            reason=reason,
-        )
-
-    def commit_testfile(self, reason=None, **kwargs):
-        reason = self._reason(reason)
-        self.write("test.txt", "test {reason}", reason=reason)
-        self(self.add_command)
-        self.commit(reason=reason, **kwargs)
-
-    def get_version(self, **kw):
-        __tracebackhide__ = True
-        from setuptools_scm import get_version
-
-        version = get_version(root=str(self.cwd), fallback_root=str(self.cwd), **kw)
-        print(version)
-        return version
-
-    @property
-    def version(self):
-        __tracebackhide__ = True
-        return self.get_version()
 
 
 @pytest.fixture(autouse=True)
@@ -106,7 +44,7 @@ def debug_mode():
 def wd(tmp_path):
     target_wd = tmp_path.resolve() / "wd"
     target_wd.mkdir()
-    return Wd(target_wd)
+    return WorkDir(target_wd)
 
 
 @pytest.fixture
@@ -117,7 +55,7 @@ def repositories_hg_git(tmp_path):
     path_git = tmp_path / "repo_git"
     path_git.mkdir()
 
-    wd = Wd(path_git)
+    wd = WorkDir(path_git)
     wd("git init")
     wd("git config user.email test@example.com")
     wd('git config user.name "a test"')
@@ -131,7 +69,7 @@ def repositories_hg_git(tmp_path):
     with open(path_hg / ".hg/hgrc", "a") as file:
         file.write("[extensions]\nhggit =\n")
 
-    wd_hg = Wd(path_hg)
+    wd_hg = WorkDir(path_hg)
     wd_hg.add_command = "hg add ."
     wd_hg.commit_command = 'hg commit -m test-{reason} -u test -d "0 0"'
